+++ AcidAnalyzer/ringBuffer.c
@@ -1,10 +1,10 @@
 #include <ringBuffer.h>
 
-void initializeBufferWithChunksSize(struct RingBuffer *rb, int numberOfBuffers, int size, int elementSize, int chunkSize)
+void initializeBufferWithChunksSize(struct RingBuffer *rb, int64_t numberOfBuffers, int64_t size, int64_t elementSize, int64_t chunkSize)
 {
     if (rb->buffers != NULL)
     {
-        for (int i = 0; i < rb->numberOfBuffers; i++)
+        for (int64_t i = 0; i < rb->numberOfBuffers; i++)
         {
             if (rb->buffers[i] != NULL) free(rb->buffers[i]);
         }
@@ -21,13 +21,13 @@ void initializeBufferWithChunksSize(struct RingBuffer *rb, int numberOfBuffers, 
     rb->partialRead = false;
     rb->tooMuch = 0;
     rb->tooLittle = 0;
-    for (int i = 0; i < rb->numberOfBuffers; i++)
+    for (int64_t i = 0; i < rb->numberOfBuffers; i++)
     {
         rb->buffers[i] = (uint8_t*)calloc(rb->size * rb->elementSize, sizeof(uint8_t));
     }
 }
 
-void initializeBuffer(struct RingBuffer *rb, int numberOfBuffers, int size, int elementSize)
+void initializeBuffer(struct RingBuffer *rb, int64_t numberOfBuffers, int64_t size, int64_t elementSize)
 {
     initializeBufferWithChunksSize(rb, numberOfBuffers, size, elementSize, 0);
 }
@@ -42,15 +42,15 @@ bool isBufferFull(struct RingBuffer* rb)
     return !((rb->size - rb->writeIndex + rb->readIndex - 1) % rb->size);
 }
 
-int getBufferWriteSpace(struct RingBuffer* rb)
+int64_t getBufferWriteSpace(struct RingBuffer* rb)
 {
     return (rb->size - rb->writeIndex + rb->readIndex - 1) % rb->size;
 }
 
-int increaseBufferWriteIndex(struct RingBuffer* rb, int count)
+int64_t increaseBufferWriteIndex(struct RingBuffer* rb, int64_t count)
 {
-    int space = (rb->size - rb->writeIndex + rb->readIndex - 1) % rb->size;
-    int toCopy = count > space ? space : count;
+    int64_t space = (rb->size - rb->writeIndex + rb->readIndex - 1) % rb->size;
+    int64_t toCopy = count > space ? space : count;
 
     if (count > space && !rb->partialWrite)
     {
@@ -70,17 +70,17 @@ int increaseBufferWriteIndex(struct RingBuffer* rb, int count)
     }
     else
     {
-        int firstHalfSize = rb->size - rb->writeIndex;
-        int secondHalfSize = toCopy - firstHalfSize;
+        int64_t firstHalfSize = rb->size - rb->writeIndex;
+        int64_t secondHalfSize = toCopy - firstHalfSize;
         rb->writeIndex = secondHalfSize;
     }
     return toCopy;
 }
 
-int writeBuffer(struct RingBuffer* rb, uint8_t *data, int bufferNumber, int count)
+int64_t writeBuffer(struct RingBuffer* rb, uint8_t *data, int64_t bufferNumber, int64_t count)
 {
-    int space = (rb->size - rb->writeIndex + rb->readIndex - 1) % rb->size;
-    int toCopy = count > space ? space : count;
+    int64_t space = (rb->size - rb->writeIndex + rb->readIndex - 1) % rb->size;
+    int64_t toCopy = count > space ? space : count;
     if (bufferNumber > rb->numberOfBuffers - 1)
     {
         return -123456789;
@@ -97,31 +97,31 @@ int writeBuffer(struct RingBuffer* rb, uint8_t *data, int bufferNumber, int coun
         rb->tooMuch++;
         return toCopy;
     }
-    static int writeTimes = 0;
+    static int64_t writeTimes = 0;
     if (rb->writeIndex + toCopy <= rb->size)
     {
         memcpy(&rb->buffers[bufferNumber][rb->writeIndex * rb->elementSize], &data[0], toCopy * rb->elementSize);
     }
     else
     {
-        int firstHalfSize = rb->size - rb->writeIndex;
+        int64_t firstHalfSize = rb->size - rb->writeIndex;
         memcpy(&rb->buffers[bufferNumber][rb->writeIndex * rb->elementSize], &data[0], firstHalfSize * rb->elementSize);
 
-        int secondHalfSize = toCopy - firstHalfSize;
+        int64_t secondHalfSize = toCopy - firstHalfSize;
         memcpy(&rb->buffers[bufferNumber][0], &data[firstHalfSize * rb->elementSize], secondHalfSize * rb->elementSize);
     }
     return toCopy;
 }
 
-int getBufferReadSpace(struct RingBuffer* rb)
+int64_t getBufferReadSpace(struct RingBuffer* rb)
 {
     return (rb->size + rb->writeIndex - rb->readIndex) % rb->size - rb->chunkSize;
 }
 
-int increaseBufferReadIndex(struct RingBuffer *rb, int count)
+int64_t increaseBufferReadIndex(struct RingBuffer *rb, int64_t count)
 {
-    int available = (rb->size + rb->writeIndex - rb->readIndex) % rb->size - rb->chunkSize;
-    int toRead = count > available ? available : count;
+    int64_t available = (rb->size + rb->writeIndex - rb->readIndex) % rb->size - rb->chunkSize;
+    int64_t toRead = count > available ? available : count;
 
     if (count > available && !rb->partialRead)  // rb->partialRead will cause problems. perhaps i whould deprecate it
     {
@@ -135,24 +135,24 @@ int increaseBufferReadIndex(struct RingBuffer *rb, int count)
         return toRead;
     }
 
-    static int readTimes = 0;
+    static int64_t readTimes = 0;
     if (rb->readIndex + toRead <= rb->size)
     {
         rb->readIndex += toRead;
     }
     else
     {
-        int firstHalfSize = rb->size - rb->readIndex;
-        int secondHalfSize = toRead - firstHalfSize;
+        int64_t firstHalfSize = rb->size - rb->readIndex;
+        int64_t secondHalfSize = toRead - firstHalfSize;
         rb->readIndex = secondHalfSize;
     }
     return toRead;
 }
 
-int readBuffer(struct RingBuffer *rb, uint8_t *destination, int bufferNumber, int count)
+int64_t readBuffer(struct RingBuffer *rb, uint8_t *destination, int64_t bufferNumber, int64_t count)
 {
-    int available = (rb->size + rb->writeIndex - rb->readIndex) % rb->size - rb->chunkSize;
-    int toRead = count > available ? available : count;
+    int64_t available = (rb->size + rb->writeIndex - rb->readIndex) % rb->size - rb->chunkSize;
+    int64_t toRead = count > available ? available : count;
 
     if (bufferNumber > rb->numberOfBuffers - 1)
     {
@@ -171,27 +171,27 @@ int readBuffer(struct RingBuffer *rb, uint8_t *destination, int bufferNumber, in
         return toRead;
     }
 
-    static int readTimes = 0;
+    static int64_t readTimes = 0;
     if (rb->readIndex + toRead <= rb->size)
     {
         memcpy(&destination[0], &rb->buffers[bufferNumber][rb->readIndex * rb->elementSize], toRead * rb->elementSize);
     }
     else
     {
-        int firstHalfSize = rb->size - rb->readIndex;
+        int64_t firstHalfSize = rb->size - rb->readIndex;
         memcpy(&destination[0], &rb->buffers[bufferNumber][rb->readIndex * rb->elementSize], firstHalfSize * rb->elementSize);
 
-        int secondHalfSize = toRead - firstHalfSize;
+        int64_t secondHalfSize = toRead - firstHalfSize;
         memcpy(&destination[firstHalfSize * rb->elementSize], &rb->buffers[bufferNumber][0], secondHalfSize * rb->elementSize);
 
     }
     return toRead;
 }
 
-int readChunkFromBuffer(struct RingBuffer *rb, uint8_t *destination, int bufferNumber)
+int64_t readChunkFromBuffer(struct RingBuffer *rb, uint8_t *destination, int64_t bufferNumber)
 {
-    int available = (rb->size + rb->writeIndex - rb->readIndex) % rb->size;
-    int toRead = rb->chunkSize > available ? available : rb->chunkSize;
+    int64_t available = (rb->size + rb->writeIndex - rb->readIndex) % rb->size;
+    int64_t toRead = rb->chunkSize > available ? available : rb->chunkSize;
 
     if (bufferNumber > rb->numberOfBuffers - 1)
     {
@@ -210,17 +210,17 @@ int readChunkFromBuffer(struct RingBuffer *rb, uint8_t *destination, int bufferN
         return toRead;
     }
 
-    static int readTimes = 0;
+    static int64_t readTimes = 0;
     if (rb->readIndex + toRead <= rb->size)
     {
         memcpy(&destination[0], &rb->buffers[bufferNumber][rb->readIndex * rb->elementSize], toRead * rb->elementSize);
     }
     else
     {
-        int firstHalfSize = rb->size - rb->readIndex;
+        int64_t firstHalfSize = rb->size - rb->readIndex;
         memcpy(&destination[0], &rb->buffers[bufferNumber][rb->readIndex * rb->elementSize], firstHalfSize * rb->elementSize);
 
-        int secondHalfSize = toRead - firstHalfSize;
+        int64_t secondHalfSize = toRead - firstHalfSize;
         memcpy(&destination[firstHalfSize * rb->elementSize], &rb->buffers[bufferNumber][0], secondHalfSize * rb->elementSize);
 
     }
@@ -231,7 +231,7 @@ void freeBuffer(struct RingBuffer* rb)
 {
     if (rb->buffers != NULL)
     {    
-        for (int i = 0; i < rb->numberOfBuffers; i++)
+        for (int64_t i = 0; i < rb->numberOfBuffers; i++)
         {
             if (rb->buffers[i] != NULL) free(rb->buffers[i]);
         }
